# openapiv2/calculator.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "calculator.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CalculatorService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/calculator/add": {
      "post": {
        "operationId": "CalculatorService_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/calculatorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calculatorRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    },
    "/v1/calculator/divide": {
      "post": {
        "operationId": "CalculatorService_Divide",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/calculatorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calculatorRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    },
    "/v1/calculator/multiply": {
      "post": {
        "operationId": "CalculatorService_Multiply",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/calculatorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calculatorRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    },
    "/v1/calculator/sub": {
      "post": {
        "operationId": "CalculatorService_Sub",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/calculatorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calculatorRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    }
  },
  "definitions": {
    "calculatorRequest": {
      "type": "object",
      "properties": {
        "num1": {
          "type": "string",
          "format": "int64"
        },
        "num2": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "calculatorResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}